requirements.txt
faker>=24.0.0
jinja2>=3.1.0
python-dotenv>=1.0.0
requests>=2.31.0
snowflake-connector-python>=3.7.0
thoughtspot-tml>=2.0.0
